cup占比，内存大小

+++ testdata/src/main/java/com/appframe/testdata/entity/Performance.java
@@ -3,7 +3,7 @@ package com.appframe.testdata.entity;
 import java.io.Serializable;
 
 public class Performance implements Serializable {
-    public int memoryUse;
+    public double memoryUse;
     public double cpuUse;
     public long recordTime;
 }
